Huge overhaul
changed damage and hp to decimal values for player and all damage sources fully finished armored skeleton sword and shield

+++ FS Unity FPS Prototype/FS FPS Project 2024/Assets/Scripts/Enemy scripts/MiniSkeleRedMageAI.cs
@@ -13,7 +13,7 @@ public class MiniSkeleRedMageAI : MonoBehaviour, IDamage
     [SerializeField] Renderer arms;
     [SerializeField] Renderer legs;
     [SerializeField] Animator anim;
-    [SerializeField] int HP;
+    [SerializeField] float HP;
     [SerializeField] int pointsToGain;
     [SerializeField] GameObject bullet;
     [SerializeField] float shootRate;
@@ -110,7 +110,7 @@ public class MiniSkeleRedMageAI : MonoBehaviour, IDamage
     }
 
     // Take Damage AI added by Matt
-    public void TakeDamage(int damage)
+    public void TakeDamage(float damage)
     {
         HP -= damage;
         StartCoroutine(FlashRed());
